requirements.txt
openai
pandas
jupyter
python-dotenv
